backend_jwt/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt>=5.0.0",
    "fastapi>=0.117.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.37.0",
    "slowapi>=0.1.9",
    "pydantic[email]>=2.0.0",
    "python-multipart>=0.0.6",
    "pyjwt>=2.10.1",
    "boto3>=1.34.0",
    "google-auth>=2.41.1",
    "google-auth-oauthlib>=1.2.2",
    "google-auth-httplib2>=0.2.0",
    "python-dotenv>=1.1.1",
]
